feat: NAG-67 create ActivatedGranolaMFC component
