+++ src/components/ProductShowcase.jsx
@@ -228,10 +228,3 @@ export default function ProductShowcase({
         </div>
     );
 }
-
-// TODO: Edit this component to take the accordions and product description into the props
-// Make another component <ActivatedGranolaMFCProductShowcase /> that uses this one and passes in the right props
-// Update the src/app/page.js file to use the new <ActivatedGranolaMFCProductShowcase /> component instead of this one directly
-// Update the src/components/ActivatedGranolaMFCProductShowcase.jsx file to use the new <ActivatedGranolaMFCProductShowcase /> component instead of this one directly
-// Ensure all necessary imports are added in the files where this component is used
-// Test the changes to ensure everything works as expected
